Failsafe support by logging data which can not be uploaded

+++ pi-php-scripts/dylos.php
@@ -7,7 +7,7 @@ class DylosMonitor
 	function __construct()
 	{
 		$this->init();
-	}
+	}	
 
 	function init()
 	{
@@ -26,29 +26,60 @@ class DylosMonitor
 		$serial->deviceOpen();
 	}
 
+	function restartDylos()
+	{
+		if (isset(DylosReaderConf::$powerRelayGPIO))
+		{
+			$gpio = DylosReaderConf::$powerRelayGPIO;
+			/* Try to switch the relay on and off */
+			print("Switching dylos off\n");
+			exec("echo 1 > /sys/class/gpio/gpio${gpio}/value");
+			sleep(20);
+			print("Switching dylos on\n");
+			exec("echo 0 > /sys/class/gpio/gpio${gpio}/value");
+		}
+	}
+
 	function listen()
 	{
 		$s = "";
+		$this->restartDylos();
+		$noupdatecount = 0;
+		$this->retryUpload=0;
 		while (1)
 		{	
-	        $v = $this->port->readPort();
-	        $this->log("read: ->'".print_r($v,true)."'");
-	        if (!($v==false || $v==""))
-	        {
-	        	$v = trim($v);
-	        	$v = explode(",",$v);
-	        	if (count($v)==2)
+			$noupdatecount++;
+	        	$v = $this->port->readPort();
+	        	$this->log("read: ->'".print_r($v,true)."'");
+	        	if (!($v==false || $v==""))
 	        	{
-	        		$this->upload($v[0],$v[1]);
-	        	}
-		    }
-		    sleep(30);
-	    }
+	        		$v = trim($v);
+	        		$v = explode(",",$v);
+	        		if (count($v)==2 && is_numeric($v[0]) && is_numeric($v[1]))
+	        		{
+	        			$this->upload($v[0],$v[1]);
+					$noupdatecount = 0;
+	        		}
+		    	}
+			
+			/* If no update for 5 minute, restart the dylos */
+			if ($noupdatecount>3)
+			{
+				print("No update for ".($noupdatecount/2)." minutes\n");
+			}
+			if ($noupdatecount>2*5 /* 5 minutes */)
+			{
+				$this->restartDylos();
+				/* Try to switch the relay on and off */
+				$noupdatecount = 0;
+			}
+		    	sleep(30);
+	    	}
 	}
 
 	function upload( $v1, $v2 )
 	{
-		$url = "http://aqicn.info/sensor/";
+		$url = "http://sensor.aqicn.org/sensor/";
 
 		$time = time();
 		$data = array(
@@ -61,6 +92,8 @@ class DylosMonitor
         		"id"=>DylosReaderConf::$sensorId,
         		/* For server backward compatiblity */
         		"clientVersion"=>2,
+			/* Keep track of the local IP adress for debugging */
+                        "ip"=>exec("/sbin/ifconfig wlan0 | grep 'inet addr:' | cut -d: -f2"),
         		/* Memory information is sent to track memory leaks */
                 	"mem"=>memory_get_usage(), 
         		"data"=>$data,
@@ -69,7 +102,7 @@ class DylosMonitor
 		$res = Http::post($url,json_encode($post));
 		$json = json_decode($res); 
 
-		$filename = "/tmp/upload.pending.json";
+		$filename = "/tmp/upload.dylos.pending.json";
 		if (!isset($json->result) || $json->result!="ok")
 		{
 			$this->log("Postponning '".json_encode($post)."'\nServer says '$res'\n");
@@ -83,17 +116,33 @@ class DylosMonitor
 			{
 				$data= json_decode(file_get_contents($filename));
 				$post = array( 
-	        		"key"=>DylosReaderConf::$sensorKey,
-	        		"id"=>DylosReaderConf::$sensorId,
-	        		/* For server backward compatiblity */
-	        		"clientVersion"=>2,
-	        		/* Memory information is sent to track memory leaks */
-	                	"mem"=>memory_get_usage(), 
-				"data"=>$data 
+	        			"key"=>DylosReaderConf::$sensorKey,
+	        			"id"=>DylosReaderConf::$sensorId,
+	        			/* For server backward compatiblity */
+	        			"clientVersion"=>2,
+	        			/* Memory information is sent to track memory leaks */
+	                		"mem"=>memory_get_usage(), 
+					"data"=>$data 
 				);
 				$res = Http::post($url,json_encode($post));
 				$this->log("Reposting ... Server says '$res'\n");
+				$json = json_decode($res);
+				if (!isset($json->result) || $json->result!="ok") 
+				{
+					/* In case the upload does not work, try 3 times before failing */
+					$this->retryUpload ++;
+					if ($this->retryUpload<3) return;
+
+					$time = time();
+					$path = realpath(dirname(__FILE__))."/logs/";
+					mkdir($path);
+					$tmpfile = $path."errlog.dylos.failed.".$time;
+					file_put_contents($tmpfile,$res);
+					$tmpfile = $path."upload.dylos.failed.".$time;
+					file_put_contents($tmpfile,json_encode($data));
+				}
 				unlink($filename);
+				$this->retryUpload=0;
 			}
 		}
 	}
